add Statistics.md
