+++ FakeNFT/Scenes /StatisticsTab/StatisticsScreen/Model/ServerUserModel.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+struct ServerUserModel: Decodable {
+    let name: String
+    let avatar: URL
+    let description: String
+    let website: URL
+    let nfts: [String]
+    let rating: String
+    let id: String
+}
