+++ Statistics.md
@@ -1,0 +1,34 @@
+**Архитектура - MVVM**
+**Способ вёртки - код**
+**Ссылка на таск-трекер: https://github.com/users/MikhailZavyalov/projects/1/views/3?filterQuery=MikhailZavyalov **
+
+
+#**Декомпозиция эпика «Статистика»**
+
+1. Вёрстка экрана рейтинга. ETA - 3 часа
+2. Вёрстка ячейки экрана рейтинга. ETA - 1.5 часа
+3. Запрос данных рейтинга с сервера. ETA - 8 часов
+4. Сортировка экрана рейтинга. ETA - 6 часов
+5. Вёрстка экрана инфо о пользователе. ETA - 3 часа
+6. Запрос данных о пользователе с сервера. ETA - 6 часов
+7. Вёрстка экрана "Коллекция NFT". ETA - 1.5 часа
+8. Вёрстка ячейки экрана "Коллекция NFT". ETA - 3 часа
+9. Запрос данных экрана "Коллекция NFT". ETA - 6 часов
+
+**Итого - 38 ч**
+
+#**Реализовано в 19 спринте:**
+
+1. Вёрстка экрана рейтинга. ETA - 3 часа. Факт.затраченное время - 4 ч
+2. Вёрстка ячейки экрана рейтинга. ETA - 1.5 часа. Факт.затраченное время - 3 ч
+3. Запрос данных рейтинга с сервера. ETA - 8 часов. Факт.затраченное время - 10 ч
+4. Сортировка экрана рейтинга. ETA - 6 часов. Факт.затраченное время - 9 ч
+5. Вёрстка экрана инфо о пользователе. ETA - 3 часа. Факт.затраченное время - 2 ч
+6. Запрос данных о пользователе с сервера. ETA - 6 часов. Факт.затраченное время - 6 ч
+7. Вёрстка экрана "Коллекция NFT". ETA - 1.5 часа. Факт.затраченное время - 1.5 ч
+8. Вёрстка ячейки экрана "Коллекция NFT". ETA - 3 часа. Факт.затраченное время - 3 ч
+9. Запрос данных экрана "Коллекция NFT". ETA - 6 часов. Факт.затраченное время - 7 ч
+
+**Фактически затраченное время на 19 спринт - 45,5 ч, ETA - 38 ч**
+
+
